updated wirelessCOM and it's class instances in Hexclaw_Main to work properly.
Need to work on side panel for special class object instances, their status and interaction. Might need to work on fitting terminalMenu's window contents internally/intelligently with the display dimensions of the actual terminal to create a smart number of createTable cells/dimensions and interactivity.
For example, need to add support for:
 - right side display alignment of cells/buttons/widgets/info
 - automatic update of the number of columns/rows based either on fixed cell/button dimensions (if window gets smaller, intelligently remove "empty" cells or crop according to internal rules), or fixed width/height.